MyHashSet 클래스 패키지 정리
