+++ src/collection/set/myhashset/MyHashSetMainV0.java
@@ -1,4 +1,4 @@
-package collection.set;
+package collection.set.myhashset;
 
 public class MyHashSetMainV0 {
     public static void main(String[] args) {
