+++ src/collection/set/myhashset/MyHashSetV1Main.java
@@ -1,4 +1,4 @@
-package collection.set;
+package collection.set.myhashset;
 
 public class MyHashSetV1Main {
     public static void main(String[] args) {
