+++ src/collection/set/myhashset/MyHashSetV2Main.java
@@ -1,4 +1,4 @@
-package collection.set;
+package collection.set.myhashset;
 
 public class MyHashSetV2Main {
     public static void main(String[] args) {
