+++ src/collection/set/myhashset/MyHashSetV3.java
@@ -1,22 +1,22 @@
-package collection.set;
+package collection.set.myhashset;
 
 import java.util.Arrays;
 import java.util.LinkedList;
 
-public class MyHashSetV2 {
+public class MyHashSetV3<E> implements MySet<E> {
 
     static final int DEFAULT_INITIAL_CAPACITY = 16;
 
-    private LinkedList<Object>[] buckets;
+    private LinkedList<E>[] buckets;
 
     private int size = 0;
     private int capacity = DEFAULT_INITIAL_CAPACITY;
 
-    public MyHashSetV2() {
+    public MyHashSetV3() {
         initBuckets();
     }
 
-    public MyHashSetV2(int capacity) {
+    public MyHashSetV3(int capacity) {
         this.capacity = capacity;
         initBuckets();
     }
@@ -28,9 +28,9 @@ public class MyHashSetV2 {
         }
     }
 
-    public boolean add(Object value) {
+    public boolean add(E value) {
         int hashIndex = hashIndex(value);
-        LinkedList<Object> bucket = buckets[hashIndex];
+        LinkedList<E> bucket = buckets[hashIndex];
 
         if (bucket.contains(value)) {
             return false;
@@ -42,16 +42,16 @@ public class MyHashSetV2 {
         return true;
     }
 
-    public boolean contains(Object searchValue) {
+    public boolean contains(E searchValue) {
         int hashIndex = hashIndex(searchValue);
-        LinkedList<Object> bucket = buckets[hashIndex];
+        LinkedList<E> bucket = buckets[hashIndex];
 
         return bucket.contains(searchValue);
     }
 
-    public boolean remove(Object value) {
+    public boolean remove(E value) {
         int hashIndex = hashIndex(value);
-        LinkedList<Object> bucket = buckets[hashIndex];
+        LinkedList<E> bucket = buckets[hashIndex];
 
         boolean result = bucket.remove(value);
         if (result) {
@@ -60,7 +60,7 @@ public class MyHashSetV2 {
         } else return false;
     }
 
-    private int hashIndex(Object value) {
+    private int hashIndex(E value) {
         return Math.abs(value.hashCode()) % capacity;
     }
 
@@ -70,7 +70,7 @@ public class MyHashSetV2 {
 
     @Override
     public String toString() {
-        return "MyHashSetV2{" +
+        return "MyHashSetV3{" +
                 "buckets=" + Arrays.toString(buckets) +
                 ", size=" + size +
                 ", capacity=" + capacity +
