+++ src/collection/set/myhashset/MyHashSetV3Main.java
@@ -1,4 +1,4 @@
-package collection.set;
+package collection.set.myhashset;
 
 public class MyHashSetV3Main {
     public static void main(String[] args) {
